+++ Exodus/src/utopia/exodus/rest/resource/ExodusResources.scala
@@ -15,14 +15,17 @@ object ExodusResources
 	// ATTRIBUTES	------------------------------
 	
 	/**
-	  * All publicly available resources
+	  * All publicly available resources (may be customized)
 	  */
-	val public = Vector(DescriptionRolesNode, LanguagesNode, LanguageFamiliaritiesNode, RolesNode, TasksNode)
-	
+	val publicDescriptions = Vector(DescriptionRolesNode, LanguagesNode, LanguageFamiliaritiesNode)
+	/**
+	 * Resources that are available publicly or with some other customizable authorization
+	 */
+	val customAuthorized = Vector(UsersNode)
 	/**
 	  * All resources that require authorization for at least some features
 	  */
-	val authorized = Vector(DevicesNode, OrganizationsNode, UsersNode)
+	val authorized = Vector(DevicesNode, OrganizationsNode, RolesNode, TasksNode)
 	
 	
 	// COMPUTED	---------------------------------
@@ -30,5 +33,21 @@ object ExodusResources
 	/**
 	  * All root level resources introduced in the Exodus project
 	  */
-	def all = public ++ authorized
+	@deprecated("Please use .default instead", "v1.0.1")
+	def all = default
+	/**
+	 * @return All Exodus resources. Resources that are not behind a session authorization don't use any authorization,
+	 *         which may be a security risk. Consider using apply(Boolean) instead or creating a custom set of
+	 *         resources with publicDescriptions, customAuthorized and authorized
+	 */
+	def default = apply(useApiKey = false)
+	
+	/**
+	 * @param useApiKey Whether api key authorization should be used for resources that would otherwise be
+	 *                  publicly available
+	 * @return All Exodus resources
+	 */
+	def apply(useApiKey: Boolean) =
+		publicDescriptions.map { factory => if (useApiKey) factory.forApiKey else factory.public } ++
+			customAuthorized.map { factory => if (useApiKey) factory.forApiKey else factory.public } ++ authorized
 }